refactor: 가위바위보, 묵찌빠 구조체화
